Fix dispatch issues for actions pre windowReady
Partially solves #10866
A second issue remains for #10866 where there's a race condition on the
tab index that is used.  This can exist without that though.

Browser process now registers the process right away and callbacks are registered from the window store.
It also waits until the window is ready before dispatching now.